backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
pydantic>=2.10.0
email-validator>=2.0.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
